requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
pyyaml>=6.0
loguru>=0.7.0
pytest>=7.4.0
